style changes
refactored QueryViewSidebar by moving filter stuff to new FilterPanel component
styled CreateSnapshotPanel to take up the entire height of the sidebar
