+++ src/components/dataset/query/sidebar/panels/CreateSnapshotPanel.jsx
@@ -5,27 +5,31 @@ import { withStyles } from '@material-ui/core/styles';
 import { actions } from 'react-redux-form';
 import { connect } from 'react-redux';
 
-import { Button, TextField } from '@material-ui/core';
+import { Button, TextField, Typography } from '@material-ui/core';
 import CreateSnapshotDropdown from '../CreateSnapshotDropdown';
 import ShareSnapshot from './ShareSnapshot';
 
 const styles = (theme) => ({
+  root: {
+    margin: theme.spacing(1),
+    display: 'grid',
+    gridTemplateRows: 'calc(100vh - 125px) 100px',
+  },
+  rowOne: {
+    gridRowStart: 1,
+    gridRowEnd: 2,
+    overflowY: 'auto',
+    overflowX: 'hidden',
+    padding: theme.spacing(1),
+  },
+  rowTwo: {
+    gridRowStart: 2,
+    gridRowEnd: 3,
+  },
   buttonContainer: {
     display: 'flex',
-    justifyContent: 'space-between',
-  },
-  saveButtonContainer: {
-    backgroundColor: theme.palette.primary.light,
-    marginBottom: theme.spacing(1),
-    padding: theme.spacing(1),
-    marginTop: theme.spacing(1),
-  },
-  saveButton: {
-    marginTop: theme.spacing(1),
-    alignSelf: 'flex-start',
-  },
-  cancelButton: {
-    alignSelf: 'flex-end',
+    justifyContent: 'flex-end',
+    paddingTop: theme.spacing(1),
   },
   textField: {
     backgroundColor: theme.palette.common.white,
@@ -75,50 +79,48 @@ export class CreateSnapshotPanel extends React.PureComponent {
     const { classes, dataset, handleCreateSnapshot } = this.props;
     const { name, description, assetName } = this.state;
     return (
-      <div className={clsx(classes.rowTwo, classes.saveButtonContainer)}>
-        <TextField
-          id="snapshotName"
-          label="Name"
-          variant="outlined"
-          size="small"
-          fullWidth
-          className={classes.textField}
-          onChange={(event) => this.setState({ name: event.target.value })}
-          value={name}
-        />
-        <TextField
-          id="snapshotDescription"
-          label="Description"
-          variant="outlined"
-          size="small"
-          multiline
-          rows={3}
-          fullWidth
-          className={classes.textField}
-          onChange={(event) => this.setState({ description: event.target.value })}
-          value={description}
-        />
-        {/* TODO: decide what to do when there's only one asset */}
-        <CreateSnapshotDropdown
-          options={dataset.schema.assets}
-          onSelectedItem={this.handleSelectAsset}
-        />
-        <div className={classes.buttonContainer}>
-          <Button
-            variant="contained"
-            className={clsx(classes.saveButton, { [classes.hide]: !open })}
-            onClick={() => handleCreateSnapshot(false)}
-          >
-            Cancel
-          </Button>
-          <Button
-            variant="contained"
-            className={clsx(classes.cancelButton, { [classes.hide]: !open })}
-            onClick={this.saveNameAndDescription}
-            disabled={assetName === '' || name === ''}
-          >
-            Next
-          </Button>
+      <div className={classes.root}>
+        <div className={classes.rowOne}>
+          <Typography variant="h6">Add Details</Typography>
+          <Typography variant="subtitle2">Snapshot Name</Typography>
+          <TextField
+            id="snapshotName"
+            variant="outlined"
+            size="small"
+            fullWidth
+            className={classes.textField}
+            onChange={(event) => this.setState({ name: event.target.value })}
+            value={name}
+          />
+          <Typography variant="subtitle2">Description</Typography>
+          <TextField
+            id="snapshotDescription"
+            variant="outlined"
+            size="small"
+            multiline
+            rows={10}
+            fullWidth
+            className={classes.textField}
+            onChange={(event) => this.setState({ description: event.target.value })}
+            value={description}
+          />
+          {/* TODO: decide what to do when there's only one asset */}
+          <CreateSnapshotDropdown
+            options={dataset.schema.assets}
+            onSelectedItem={this.handleSelectAsset}
+          />
+        </div>
+        <div className={classes.rowTwo}>
+          <div className={classes.buttonContainer}>
+            <Button onClick={() => handleCreateSnapshot(false)}>Cancel</Button>
+            <Button
+              variant="contained"
+              onClick={this.saveNameAndDescription}
+              disabled={assetName === '' || name === ''}
+            >
+              Next
+            </Button>
+          </div>
         </div>
       </div>
     );
@@ -129,6 +131,7 @@ function mapStateToProps(state) {
   return {
     snapshot: state.snapshot,
     snapshots: state.snapshots,
+    dataset: state.datasets.dataset,
   };
 }
